Add max score, fix float damage not working

+++ src/PitPat/Assets/Scripts/PlayerController.cs
@@ -64,7 +64,7 @@ public class PlayerController : MonoBehaviour
     /// NOTE: AttackGrid must be [7,7]
     /// </summary>
     /// <param name="attackGrid"></param>
-    public void Attack(float[,] attackGrid,int damage)
+    public void Attack(float[,] attackGrid,float damage)
     {
         if(BeatSpawner.canBePressed==true && beatHit == false)
         {
